move: Location, LocationData and Token to brack-common
