+++ crates/brack-common/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "brack-common"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde = "1.0.219"
+serde_json = "1.0.140"
